+++ Project/Minimax.h
@@ -11,11 +11,11 @@ namespace twixt {
 	class Minimax
 	{
 	private:
-		int evaluate(std::pair<twixt::Dot*, twixt::Dot*> bridgeToEvaluate);
+		uint16_t evaluate(std::pair<twixt::Dot*, twixt::Dot*> bridgeToEvaluate);
 		std::pair<twixt::Dot*, twixt::Dot*> minimax(Dot::DotStatus status);
 		void canBlock(Dot* centralDot);
 		Dot* blockOpponent(Dot* centralDot, Dot* firstOpponentDot, Dot* secondOpponentDot);
-		int longestPath(Dot* dot);
+		uint16_t longestPath(Dot* dot);
 		void scorePossibleBridges(Dot* dot);
 
 		Board *copyOfBoard;
@@ -35,7 +35,7 @@ namespace twixt {
 				return hash1 ^ hash2;
 			}
 		};
-		std::unordered_map<std::pair<Dot*, Dot*>, int, PairDotHash> mapBridges;
+		std::unordered_map<std::pair<Dot*, Dot*>, uint16_t, PairDotHash> mapBridges;
 
 
 	public:
